Select checks by check group

Returns the names of all checks that belong to the given group(s). This makes it easy to run or inspect a specific category of checks without knowing individual check names.

Usage

checks_by_group(...)

Arguments

...

Group names as character strings. Use all_check_groups() to see available names.

Value

Character vector of check names

Examples

# run only DESCRIPTION and namespace checks
checks_by_group("description", "namespace")
#>  [1] "no_obsolete_deps"                   "complexity_unused_internal"        
#>  [3] "no_description_depends"             "no_description_date"               
#>  [5] "description_url"                    "description_not_start_with_package"
#>  [7] "description_urls_in_angle_brackets" "description_doi_format"            
#>  [9] "description_urls_not_http"          "no_description_duplicate_deps"     
#> [11] "description_valid_roles"            "description_pkgname_single_quoted" 
#> [13] "description_bugreports"             "no_import_package_as_a_whole"      
#> [15] "no_export_pattern"                  "rd_has_examples"                   
#> [17] "rd_has_return"                      "reverse_dependencies"              

# see what the lintr group covers
checks_by_group("lintr")
#>  [1] "lintr_assignment_linter"               
#>  [2] "lintr_line_length_linter"              
#>  [3] "lintr_semicolon_linter"                
#>  [4] "lintr_attach_detach_linter"            
#>  [5] "lintr_setwd_linter"                    
#>  [6] "lintr_sapply_linter"                   
#>  [7] "lintr_library_require_linter"          
#>  [8] "installed_packages_linter"             
#>  [9] "lintr_seq_linter"                      
#> [10] "lintr_any_duplicated_linter"           
#> [11] "lintr_any_is_na_linter"                
#> [12] "lintr_class_equals_linter"             
#> [13] "lintr_condition_message_linter"        
#> [14] "lintr_duplicate_argument_linter"       
#> [15] "lintr_equals_na_linter"                
#> [16] "lintr_fixed_regex_linter"              
#> [17] "lintr_for_loop_index_linter"           
#> [18] "lintr_length_test_linter"              
#> [19] "lintr_matrix_apply_linter"             
#> [20] "lintr_missing_argument_linter"         
#> [21] "lintr_nrow_subset_linter"              
#> [22] "lintr_redundant_equals_linter"         
#> [23] "lintr_redundant_ifelse_linter"         
#> [24] "lintr_regex_subset_linter"             
#> [25] "lintr_sort_linter"                     
#> [26] "lintr_system_file_linter"              
#> [27] "lintr_terminal_close_linter"           
#> [28] "lintr_which_grepl_linter"              
#> [29] "lintr_boolean_arithmetic_linter"       
#> [30] "lintr_comparison_negation_linter"      
#> [31] "lintr_consecutive_assertion_linter"    
#> [32] "lintr_if_not_else_linter"              
#> [33] "lintr_ifelse_censor_linter"            
#> [34] "lintr_implicit_assignment_linter"      
#> [35] "lintr_inner_combine_linter"            
#> [36] "lintr_length_levels_linter"            
#> [37] "lintr_literal_coercion_linter"         
#> [38] "lintr_nested_ifelse_linter"            
#> [39] "lintr_nested_pipe_linter"              
#> [40] "lintr_numeric_leading_zero_linter"     
#> [41] "lintr_outer_negation_linter"           
#> [42] "lintr_paste_linter"                    
#> [43] "lintr_scalar_in_linter"                
#> [44] "lintr_undesirable_operator_linter"     
#> [45] "lintr_unnecessary_concatenation_linter"
#> [46] "lintr_unnecessary_lambda_linter"       
#> [47] "lintr_unreachable_code_linter"         
#> [48] "lintr_conjunct_test_linter"            
#> [49] "lintr_expect_comparison_linter"        
#> [50] "lintr_expect_identical_linter"         
#> [51] "lintr_expect_length_linter"            
#> [52] "lintr_expect_named_linter"             
#> [53] "lintr_expect_not_linter"               
#> [54] "lintr_expect_null_linter"              
#> [55] "lintr_expect_s3_class_linter"          
#> [56] "lintr_expect_s4_class_linter"          
#> [57] "lintr_expect_true_false_linter"        
#> [58] "lintr_expect_type_linter"              
# See all checks by group:
lapply(all_check_groups(), checks_by_group)
#> [[1]]
#> [1] "covr"
#> 
#> [[2]]
#> [1] "cyclocomp"
#> 
#> [[3]]
#>  [1] "no_obsolete_deps"                   "no_description_depends"            
#>  [3] "no_description_date"                "description_url"                   
#>  [5] "description_not_start_with_package" "description_urls_in_angle_brackets"
#>  [7] "description_doi_format"             "description_urls_not_http"         
#>  [9] "no_description_duplicate_deps"      "description_valid_roles"           
#> [11] "description_pkgname_single_quoted"  "description_bugreports"            
#> [13] "reverse_dependencies"              
#> 
#> [[4]]
#>  [1] "lintr_assignment_linter"               
#>  [2] "lintr_line_length_linter"              
#>  [3] "lintr_semicolon_linter"                
#>  [4] "lintr_attach_detach_linter"            
#>  [5] "lintr_setwd_linter"                    
#>  [6] "lintr_sapply_linter"                   
#>  [7] "lintr_library_require_linter"          
#>  [8] "installed_packages_linter"             
#>  [9] "lintr_seq_linter"                      
#> [10] "lintr_any_duplicated_linter"           
#> [11] "lintr_any_is_na_linter"                
#> [12] "lintr_class_equals_linter"             
#> [13] "lintr_condition_message_linter"        
#> [14] "lintr_duplicate_argument_linter"       
#> [15] "lintr_equals_na_linter"                
#> [16] "lintr_fixed_regex_linter"              
#> [17] "lintr_for_loop_index_linter"           
#> [18] "lintr_length_test_linter"              
#> [19] "lintr_matrix_apply_linter"             
#> [20] "lintr_missing_argument_linter"         
#> [21] "lintr_nrow_subset_linter"              
#> [22] "lintr_redundant_equals_linter"         
#> [23] "lintr_redundant_ifelse_linter"         
#> [24] "lintr_regex_subset_linter"             
#> [25] "lintr_sort_linter"                     
#> [26] "lintr_system_file_linter"              
#> [27] "lintr_terminal_close_linter"           
#> [28] "lintr_which_grepl_linter"              
#> [29] "lintr_boolean_arithmetic_linter"       
#> [30] "lintr_comparison_negation_linter"      
#> [31] "lintr_consecutive_assertion_linter"    
#> [32] "lintr_if_not_else_linter"              
#> [33] "lintr_ifelse_censor_linter"            
#> [34] "lintr_implicit_assignment_linter"      
#> [35] "lintr_inner_combine_linter"            
#> [36] "lintr_length_levels_linter"            
#> [37] "lintr_literal_coercion_linter"         
#> [38] "lintr_nested_ifelse_linter"            
#> [39] "lintr_nested_pipe_linter"              
#> [40] "lintr_numeric_leading_zero_linter"     
#> [41] "lintr_outer_negation_linter"           
#> [42] "lintr_paste_linter"                    
#> [43] "lintr_scalar_in_linter"                
#> [44] "lintr_undesirable_operator_linter"     
#> [45] "lintr_unnecessary_concatenation_linter"
#> [46] "lintr_unnecessary_lambda_linter"       
#> [47] "lintr_unreachable_code_linter"         
#> [48] "lintr_conjunct_test_linter"            
#> [49] "lintr_expect_comparison_linter"        
#> [50] "lintr_expect_identical_linter"         
#> [51] "lintr_expect_length_linter"            
#> [52] "lintr_expect_named_linter"             
#> [53] "lintr_expect_not_linter"               
#> [54] "lintr_expect_null_linter"              
#> [55] "lintr_expect_s3_class_linter"          
#> [56] "lintr_expect_s4_class_linter"          
#> [57] "lintr_expect_true_false_linter"        
#> [58] "lintr_expect_type_linter"              
#> 
#> [[5]]
#> [1] "complexity_unused_internal"   "no_import_package_as_a_whole"
#> [3] "no_export_pattern"            "rd_has_examples"             
#> [5] "rd_has_return"               
#> 
#> [[6]]
#>   [1] "rcmdcheck_package_directory_exists"                 
#>   [2] "rcmdcheck_vignette_leftovers"                       
#>   [3] "rcmdcheck_qpdf_present_for_checking"                
#>   [4] "rcmdcheck_vignettes_at_right_place"                 
#>   [5] "rcmdcheck_portable_file_names"                      
#>   [6] "rcmdcheck_duplicate_file_names"                     
#>   [7] "rcmdcheck_ascii_file_names"                         
#>   [8] "rcmdcheck_portable_paths"                           
#>   [9] "rcmdcheck_proper_permissions"                       
#>  [10] "rcmdcheck_executable_files"                         
#>  [11] "rcmdcheck_description_meta"                         
#>  [12] "rcmdcheck_malformed_title_or_description"           
#>  [13] "rcmdcheck_portable_description_encoding"            
#>  [14] "rcmdcheck_description_encoding_known"               
#>  [15] "rcmdcheck_description_ascii_tags"                   
#>  [16] "rcmdcheck_description_ascii_fields"                 
#>  [17] "rcmdcheck_standardizable_license"                   
#>  [18] "rcmdcheck_license_file_pointers"                    
#>  [19] "rcmdcheck_non_standard_license"                     
#>  [20] "rcmdcheck_deprecated_license"                       
#>  [21] "rcmdcheck_not_permitted_license_restrictions"       
#>  [22] "rcmdcheck_template_licenses"                        
#>  [23] "rcmdcheck_authors_at_r_field"                       
#>  [24] "rcmdcheck_valid_author_in_authors_at_r"             
#>  [25] "rcmdcheck_author_in_authors_at_r"                   
#>  [26] "rcmdcheck_authors_at_r_valid_roles"                 
#>  [27] "rcmdcheck_maintainer_in_authors_at_r"               
#>  [28] "rcmdcheck_single_maintainer_in_authors_at_r"        
#>  [29] "rcmdcheck_valid_maintainer_in_authors_at_t"         
#>  [30] "rcmdcheck_stale_author"                             
#>  [31] "rcmdcheck_stale_maintainer"                         
#>  [32] "rcmdcheck_needscompilation_field_valid"             
#>  [33] "rcmdcheck_multiple_dependency_modes"                
#>  [34] "rcmdcheck_unused_linkingto"                         
#>  [35] "rcmdcheck_versioned_linkingto"                      
#>  [36] "rcmdcheck_linkingto_without_compiled_code"          
#>  [37] "rcmdcheck_broken_vignettes_1"                       
#>  [38] "rcmdcheck_broken_vignettes_2"                       
#>  [39] "rcmdcheck_defunct_top_level"                        
#>  [40] "rcmdcheck_valid_readme_and_news"                    
#>  [41] "rcmdcheck_stale_license_file"                       
#>  [42] "rcmdcheck_stale_inst_license_file"                  
#>  [43] "rcmdcheck_non_standard_top_level_files"             
#>  [44] "rcmdcheck_copyright_file_in_inst"                   
#>  [45] "rcmdcheck_authors_in_inst"                          
#>  [46] "rcmdcheck_leftover_files"                           
#>  [47] "rcmdcheck_empty_index_file"                         
#>  [48] "rcmdcheck_empty_demo_index_file"                    
#>  [49] "rcmdcheck_proper_demo_index_file"                   
#>  [50] "rcmdcheck_demos_are_indexed"                        
#>  [51] "rcmdcheck_stale_demos_in_index"                     
#>  [52] "rcmdcheck_vignette_index"                           
#>  [53] "rcmdcheck_has_r_source_files"                       
#>  [54] "rcmdcheck_r_directory_uppercase"                    
#>  [55] "rcmdcheck_man_directory_lowercase"                  
#>  [56] "rcmdcheck_data_directory_lowercase"                 
#>  [57] "rcmdcheck_rcheck_directories"                       
#>  [58] "rcmdcheck_leftover_dvi_pdf_build_directories"       
#>  [59] "rcmdcheck_leftover_vs_directories"                  
#>  [60] "rcmdcheck_invalid_file_names"                       
#>  [61] "rcmdcheck_empty_data_directory"                     
#>  [62] "rcmdcheck_empty_demo_directory"                     
#>  [63] "rcmdcheck_demos_are_ascii"                          
#>  [64] "rcmdcheck_demos_have_valid_code"                    
#>  [65] "rcmdcheck_empty_exec_directory"                     
#>  [66] "rcmdcheck_empty_inst_directory"                     
#>  [67] "rcmdcheck_src_without_sources"                      
#>  [68] "rcmdcheck_inst_interference"                        
#>  [69] "rcmdcheck_news_rd_is_valid"                         
#>  [70] "rcmdcheck_citation_file_is_valid"                   
#>  [71] "rcmdcheck_citation_file_at_standard_place"          
#>  [72] "rcmdcheck_r_files_are_ascii"                        
#>  [73] "rcmdcheck_errors_in_r_files"                        
#>  [74] "rcmdcheck_warnings_in_r_files"                      
#>  [75] "rcmdcheck_undeclared_imports"                       
#>  [76] "rcmdcheck_undeclared_library_require"               
#>  [77] "rcmdcheck_undeclared_loadnamespace_requirenamespace"
#>  [78] "rcmdcheck_library_require_to_attached"              
#>  [79] "rcmdcheck_library_require_in_package_code"          
#>  [80] "rcmdcheck_imports_not_imported_from"                
#>  [81] "rcmdcheck_depends_not_imported_from"                
#>  [82] "rcmdcheck_missing_or_unexported_objects"            
#>  [83] "rcmdcheck_unneeded_triple_colon"                    
#>  [84] "rcmdcheck_triple_colon_imported_objects_exist"      
#>  [85] "rcmdcheck_unexported_base_objects_imported"         
#>  [86] "rcmdcheck_unexported_objects_imported"              
#>  [87] "rcmdcheck_triple_colon_to_itself"                   
#>  [88] "rcmdcheck_triple_colon_to_unknown"                  
#>  [89] "rcmdcheck_data_call_to_undeclared_package"          
#>  [90] "rcmdcheck_exported_s3_methods_are_registered"       
#>  [91] "rcmdcheck_replacement_function_arg_names"           
#>  [92] "rcmdcheck_foreign_calls_have_package_argument"      
#>  [93] "rcmdcheck_foreign_call_empty_package_argument"      
#>  [94] "rcmdcheck_foreign_call_to_base_package"             
#>  [95] "rcmdcheck_foreign_calls_to_another_package"         
#>  [96] "rcmdcheck_foreign_calls_to_unknown_package"         
#>  [97] "rcmdcheck_foreign_registration_problems"            
#>  [98] "rcmdcheck_foreign_calls_with_dup"                   
#>  [99] "rcmdcheck_loading_and_unloading"                    
#> [100] "rcmdcheck_first_lib_obsolete"                       
#> [101] "rcmdcheck_startup_function_arguments"               
#> [102] "rcmdcheck_startup_function_change_search_path"      
#> [103] "rcmdcheck_startup_function_messages"                
#> [104] "rcmdcheck_statup_function_unsafe_calls"             
#> [105] "rcmdcheck_last_lib_needs_to_be_exported"            
#> [106] "rcmdcheck_detach_function_arguments"                
#> [107] "rcmdcheck_detach_no_library_dynam_unload"           
#> [108] "rcmdcheck_unsafe_calls"                             
#> [109] "rcmdcheck_partial_argument_match"                   
#> [110] "rcmdcheck_undefined_globals"                        
#> [111] "rcmdcheck_avoid_internal_calls"                     
#> [112] "rcmdcheck_avoid_internal_calls_s4"                  
#> [113] "rcmdcheck_avoid_internal_calls_rc"                  
#> [114] "rcmdcheck_assignment_to_globalenv"                  
#> [115] "rcmdcheck_avoid_using_attach"                       
#> [116] "rcmdcheck_data_into_globalenv"                      
#> [117] "rcmdcheck_obsolete_platform_specific"               
#> [118] "rcmdcheck_obsolete_platform_specific_s4"            
#> [119] "rcmdcheck_obsolete_platform_specific_rc"            
#> [120] "rcmdcheck_deprecated_functions"                     
#> [121] "rcmdcheck_defunct_removed_functions"                
#> [122] "rcmdcheck_avoid_platform_specific_devices"          
#> [123] "rcmdcheck_rd_empty_sections"                        
#> [124] "rcmdcheck_rd_problems"                              
#> [125] "rcmdcheck_rd_duplicated_name"                       
#> [126] "rcmdcheck_rd_duplicated_alias"                      
#> [127] "rcmdcheck_rd_long_code_lines"                       
#> [128] "rcmdcheck_rd_cross_references"                      
#> [129] "rcmdcheck_missing_docs"                             
#> [130] "rcmdcheck_code_docs_mismatch"                       
#> [131] "rcmdcheck_rd_usage"                                 
#> [132] "rcmdcheck_rd_s3_usage"                              
#> [133] "rcmdcheck_rd_contents"                              
#> [134] "rcmdcheck_unstated_dependencies_in_examples"        
#> [135] "rcmdcheck_data_contents"                            
#> [136] "rcmdcheck_undeclared_non_ascii_characters_in_data"  
#> [137] "rcmdcheck_non_ascii_characters_in_data"             
#> [138] "rcmdcheck_uncompressed_data_files"                  
#> [139] "rcmdcheck_bzip2_xz_requirement"                     
#> [140] "rcmdcheck_uncompressed_sysdata"                     
#> [141] "rcmdcheck_unneeded_doc_extra"                       
#> [142] "rcmdcheck_doc_extra_licenses"                       
#> [143] "rcmdcheck_doc_tex_leftovers"                        
#> [144] "rcmdcheck_extra_files_in_docs"                      
#> [145] "rcmdcheck_extra_directories_in_docs"                
#> [146] "rcmdcheck_vignettes_in_vignettes"                   
#> [147] "rcmdcheck_knitr_leftovers"                          
#> [148] "rcmdcheck_pdf_file_sizes"                           
#> [149] "rcmdcheck_pdf_file_sizes_gs"                        
#> [150] "rcmdcheck_source_line_endings"                      
#> [151] "rcmdcheck_makefile_line_endings"                    
#> [152] "rcmdcheck_makefile_with_final_lf"                   
#> [153] "rcmdcheck_non_portable_makevars"                    
#> [154] "rcmdcheck_makevars_overriding_user_site"            
#> [155] "rcmdcheck_makevars_and_makevars_in"                 
#> [156] "rcmdcheck_gnu_make_required"                        
#> [157] "rcmdcheck_gnu_makefile_extensions"                  
#> [158] "rcmdcheck_blas_and_lapack_flags"                    
#> [159] "rcmdcheck_missing_flibs"                            
#> [160] "rcmdcheck_unsafe_calls_in_compiled_code"            
#> [161] "rcmdcheck_loading_package"                          
#> [162] "rcmdcheck_unloading_package"                        
#> [163] "rcmdcheck_namespace_can_be_loaded"                  
#> [164] "rcmdcheck_namespace_can_be_loaded_safely"           
#> [165] "rcmdcheck_namespace_can_be_unloaded"                
#> [166] "rcmdcheck_loading_when_not_on_search_path"          
#> [167] "rcmdcheck_s3_method_registration"                   
#> [168] "rcmdcheck_encoding_in_ascii_locale"                 
#> [169] "rcmdcheck_examples_run"                             
#> [170] "rcmdcheck_examples_run_without_warnings"            
#> [171] "rcmdcheck_two_many_cores_used"                      
#> [172] "rcmdcheck_can_collect_examples"                     
#> [173] "rcmdcheck_unstated_dependencies_in_tests"           
#> [174] "rcmdcheck_tests_pass"                               
#> [175] "rcmdcheck_unstated_dependencies_in_vignettes"       
#> [176] "rcmdcheck_vignette_output_present"                  
#> [177] "rcmdcheck_encoding_defined_in_vignettes"            
#> [178] "rcmdcheck_doc_makefile_uppercase"                   
#> [179] "rcmdcheck_calling_r_from_makefile"                  
#> [180] "rcmdcheck_makefile_line_endings_2"                  
#> [181] "rcmdcheck_calling_rscript_from_makefile"            
#> [182] "rcmdcheck_correct_vignette_encodings"               
#> [183] "rcmdcheck_vignettes_load_dependencies"              
#> [184] "rcmdcheck_vignettes_run"                            
#> [185] "rcmdcheck_vignettes_build"                          
#> [186] "rcmdcheck_vignettes_build_2"                        
#> [187] "rcmdcheck_can_convert_rd_to_pdf"                    
#> [188] "rcmdcheck_can_convert_rd_to_pdf_2"                  
#> [189] "rcmdcheck_pdf_without_hyperref"                     
#> [190] "rcmdcheck_executable_files_in_packages"             
#> [191] "rcmdcheck_hidden_files_and_directories"             
#> [192] "rcmdcheck_installs"                                 
#> [193] "rcmdcheck_significant_compilation_warnings"         
#> [194] "rcmdcheck_other_compilation_warnings"               
#> [195] "rcmdcheck_reasonable_installed_size"                
#> [196] "rcmdcheck_description_required_fields"              
#> [197] "rcmdcheck_package_name_portable"                    
#> [198] "rcmdcheck_description_right_case"                   
#> [199] "rcmdcheck_not_package_bundle"                       
#> [200] "rcmdcheck_cran_incoming_feasibility_1"              
#> [201] "rcmdcheck_cran_incoming_feasibility_2"              
#> [202] "rcmdcheck_cran_incoming_feasibility_3"              
#> [203] "rcmdcheck_valid_namespace"                          
#> [204] "rcmdcheck_empty_importfrom_in_namespace"            
#> [205] "rcmdcheck_too_many_s3_methods"                      
#> [206] "rcmdcheck_package_dependencies_present"             
#> [207] "rcmdcheck_if_source_package"                        
#> [208] "rcmdcheck_object_files_in_source_package"           
#> [209] "rcmdcheck_multi_arch_build_dir"                     
#> [210] "rcmdcheck_compilation_leftover_files"               
#> [211] "rcmdcheck_object_files_in_source_package_2"         
#> [212] "rcmdcheck_installed_version_included"               
#> [213] "rcmdcheck_some_code_in_docs"                        
#> 
#> [[7]]
#> [1] "rd_has_examples" "rd_has_return"  
#> 
#> [[8]]
#> [1] "reverse_dependencies"
#> 
#> [[9]]
#> [1] "roxygen2_has_export_or_nord" "roxygen2_unknown_tags"      
#> [3] "roxygen2_valid_inherit"      "roxygen2_duplicate_params"  
#> 
#> [[10]]
#> [1] "print_return_invisible"     "on_exit_has_add"           
#> [3] "complexity_function_length" "duplicate_function_bodies" 
#> 
#> [[11]]
#> [1] "has_readme"       "has_news"         "r_file_extension"
#> 
#> [[12]]
#> [1] "spelling"
#> 
#> [[13]]
#>  [1] "tidyverse_brace_linter"                    
#>  [2] "tidyverse_commas_linter"                   
#>  [3] "tidyverse_commented_code_linter"           
#>  [4] "tidyverse_equals_na_linter"                
#>  [5] "tidyverse_function_left_parentheses_linter"
#>  [6] "tidyverse_indentation_linter"              
#>  [7] "tidyverse_infix_spaces_linter"             
#>  [8] "tidyverse_object_length_linter"            
#>  [9] "tidyverse_object_name_linter"              
#> [10] "tidyverse_object_usage_linter"             
#> [11] "tidyverse_paren_body_linter"               
#> [12] "tidyverse_pipe_consistency_linter"         
#> [13] "tidyverse_pipe_continuation_linter"        
#> [14] "tidyverse_quotes_linter"                   
#> [15] "tidyverse_return_linter"                   
#> [16] "tidyverse_spaces_inside_linter"            
#> [17] "tidyverse_spaces_left_parentheses_linter"  
#> [18] "tidyverse_trailing_blank_lines_linter"     
#> [19] "tidyverse_trailing_whitespace_linter"      
#> [20] "tidyverse_vector_logic_linter"             
#> [21] "tidyverse_whitespace_linter"               
#> [22] "tidyverse_assignment_linter"               
#> [23] "tidyverse_line_length_linter"              
#> [24] "tidyverse_semicolon_linter"                
#> [25] "tidyverse_seq_linter"                      
#> [26] "tidyverse_T_and_F_symbol_linter"           
#> [27] "tidyverse_r_file_names"                    
#> [28] "tidyverse_test_file_names"                 
#> [29] "tidyverse_no_missing"                      
#> [30] "tidyverse_export_order"                    
#> 
#> [[14]]
#> [1] "urlchecker_ok"           "urlchecker_no_redirects"
#> 
#> [[15]]
#> [1] "vignette_no_rm_list" "vignette_no_setwd"  
#> 
# use directly in gp()
if (FALSE) { # \dontrun{
  gp(".", checks = checks_by_group("description", "lintr"))
} # }